Valid a brackets expression
Basic approach with the help of stack. When it is an open bracket, we
put it to the stack, when it is a close bracket, we check from the
stack and see if it is a correct one. If so, we pop the top element
from the stack. At the end. if the stack is empty. We have a balance
brackets expression